Weather Analysis V 0.0.2 - Refined compareDateTempsPerYear function
further developed compareDateTempsPerYear function in the hope that it would input the temperature on a given date in a given year. Last error stated that the truth of a Series is ambiguous, likely meaning that there is more than one location where the conditions set are true. So, I'm gonna have to refine it. But, the script looks to be in an okay spot as is. Will verify later

I should start a DevNotes text file.