Weather Analysis V 0.0.3 - Temperature Comparison Plots Done
Finalized Function to draw line plots for given locations. Added a try-catch error for missing values if it occurs in the data.

Could place the median value in there just in case, but for plotting that doesn't have much use.

Script seems stable

Need to start DevNotes for this project