Weather Analysis V 0.0.8 - Fixed Zero Division Error More details in the "devnotes" document. But bottom line, the problem wasn't tokyo. Los Angeles just hasn't had any snow in the city itself for God knows how many years
I placed a try-except statement where the min, max, range, scope values are initialized for the first loop (for i ...).
Seems to have worked. Many of the histograms have significant outliers though.
So it looks like I'll have to implement a zscore filter to narrow the scope of observed data points to those which fall within 2.5-3 standard deviations of the mean.

Hopefully that's the solution.

Enough for today. Time to complete homeowork.